transfer funds between two clients

+++ src/main/java/com/martinPluhar/Bankapplication/dto/TransferRequest.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -12,5 +14,5 @@ import lombok.NoArgsConstructor;
 public class TransferRequest {
     String sourceAccountNumber;
     String destinationAccountNumber;
-    double amount;
+    BigDecimal amount;
 }
